cell and grid system 0.2.1 b:
- new, simpler calculation of a cell. Working for every height and width of a cell

	modified:   TowerDefenceANTSPrototype2/src/basics/ANTSDriver.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSCellModel.java
	modified:   TowerDefenceANTSPrototype2/src/models/ANTSGridModel.java

+++ TowerDefenceANTSPrototype2/src/models/ANTSGridModel.java
@@ -66,7 +66,7 @@ public class ANTSGridModel extends ANTSAbstractModel implements ANTSIModel
 	//GETTERS & SETTERS//
 	/////////////////////
 	
-	public int getOffsetX()
+	public double getOffsetX()
 	{
 		return this.xOffset - ANTSCellModel.offsetCellX/2;
 	}
